+++ CoordinateCalculator/CoordinateCalculator/Pointable.swift
@@ -1,0 +1,15 @@
+//
+//  MyShape.swift
+//  CoordinateCalculator
+//
+//  Created by hw on 26/04/2019.
+//  Copyright © 2019 Codesquad Inc. All rights reserved.
+//
+
+import Foundation
+
+protocol Pointable {
+    func getDrawablePointList() -> [MyPoint]
+}
+
+
